requirements.txt
pizzoo
Pillow
requests
FlightRadar24
